# APIdocs.json
{
    "openapi":"3.0.1",
    "info":{
       "title":"OpenAPI definition",
       "version":"v0"
    },
    "servers":[
       {
          "url":"http://localhost:8080",
          "description":"Generated server url"
       }
    ],
    "paths":{
       "/rates/{currency}&{targets}":{
          "get":{
             "tags":[
                "exchange-rate-controller"
             ],
             "operationId":"getExchangeRates",
             "parameters":[
                {
                   "name":"currency",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"targets",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/JSONObject"
                         }
                      }
                   }
                }
             }
          }
       },
       "/rates/{api}/{currency}&{targets}":{
          "get":{
             "tags":[
                "exchange-rate-controller"
             ],
             "operationId":"getExchangeRates_1",
             "parameters":[
                {
                   "name":"currency",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"targets",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"api",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/JSONObject"
                         }
                      }
                   }
                }
             }
          }
       },
       "/rates/all/{currency}":{
          "get":{
             "tags":[
                "exchange-rate-controller"
             ],
             "operationId":"getAllExchangeRates",
             "parameters":[
                {
                   "name":"currency",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/JSONObject"
                         }
                      }
                   }
                }
             }
          }
       },
       "/rates/all/{api}/{currency}":{
          "get":{
             "tags":[
                "exchange-rate-controller"
             ],
             "operationId":"getAllExchangeRates_1",
             "parameters":[
                {
                   "name":"currency",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"api",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/JSONObject"
                         }
                      }
                   }
                }
             }
          }
       },
       "/convert/{from}&{to}&{amount}":{
          "get":{
             "tags":[
                "exchange-rate-controller"
             ],
             "operationId":"convertCurrency",
             "parameters":[
                {
                   "name":"from",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"to",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"amount",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"integer",
                      "format":"int32"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/JSONObject"
                         }
                      }
                   }
                }
             }
          }
       },
       "/convert/{api}/{from}&{to}&{amount}":{
          "get":{
             "tags":[
                "exchange-rate-controller"
             ],
             "operationId":"convertCurrency_1",
             "parameters":[
                {
                   "name":"from",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"to",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"amount",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"integer",
                      "format":"int32"
                   }
                },
                {
                   "name":"api",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/JSONObject"
                         }
                      }
                   }
                }
             }
          }
       },
       "/convert/multi/{from}&{targets}&{amount}":{
          "get":{
             "tags":[
                "exchange-rate-controller"
             ],
             "operationId":"convertMultiCurrency",
             "parameters":[
                {
                   "name":"from",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"targets",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"amount",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"integer",
                      "format":"int32"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/JSONObject"
                         }
                      }
                   }
                }
             }
          }
       },
       "/convert/multi/{api}/{from}&{targets}&{amount}":{
          "get":{
             "tags":[
                "exchange-rate-controller"
             ],
             "operationId":"convertMultiCurrency_1",
             "parameters":[
                {
                   "name":"from",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"targets",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                },
                {
                   "name":"amount",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"integer",
                      "format":"int32"
                   }
                },
                {
                   "name":"api",
                   "in":"path",
                   "required":true,
                   "schema":{
                      "type":"string"
                   }
                }
             ],
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "$ref":"#/components/schemas/JSONObject"
                         }
                      }
                   }
                }
             }
          }
       },
       "/":{
          "get":{
             "tags":[
                "exchange-rate-controller"
             ],
             "operationId":"index",
             "responses":{
                "200":{
                   "description":"OK",
                   "content":{
                      "*/*":{
                         "schema":{
                            "type":"string"
                         }
                      }
                   }
                }
             }
          }
       }
    },
    "components":{
       "schemas":{
          "JSONObject":{
             "type":"object",
             "properties":{
                "empty":{
                   "type":"boolean"
                }
             },
             "additionalProperties":{
                "type":"object"
             }
          }
       }
    }
 }